+++ assets/javascripts/kitten/components/meters/progress-ring/index.js
@@ -26,7 +26,7 @@ const StyledMeterCircle = styled.circle`
     getDashOffset({ r, progressValue })};
   transform: rotate(-90deg);
   transform-origin: ${({ cx, cy }) => `${pxToRem(cx)} ${pxToRem(cy)}`};
-  animation: ${rotateAnimate} ${({ speedAnimation }) => speedAnimation} ease-out;
+  animation: ${rotateAnimate} ${({ animationSpeed }) => animationSpeed} ease-out;
   
   &.k-Meters__ProgressRing--orion {
     stroke-linecap: round;
@@ -40,7 +40,7 @@ export const ProgressRing = ({
   strokeWidth,
   variant,
   className,
-  speedAnimation,
+  animationSpeed,
   ...others
 }) => {
   const circleX = width / 2
@@ -73,7 +73,7 @@ export const ProgressRing = ({
         cy={circleY}
         r={radius}
         strokeWidth={strokeWidth}
-        speedAnimation={`${speedAnimation}s`}
+        animationSpeed={`${animationSpeed}s`}
         fill="transparent"
         stroke={color}
         progressValue={progressValue}
@@ -94,7 +94,7 @@ ProgressRing.defaultProps = {
   radius: null,
   strokeWidth: 5,
   variant: 'andromeda',
-  speedAnimation: '1.4s',
+  animationSpeed: 1.4,
 }
 
 ProgressRing.propTypes = {
@@ -119,5 +119,5 @@ ProgressRing.propTypes = {
   */
   strokeWidth: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
   variant: PropTypes.oneOf(['andromeda', 'orion']),
-  speedAnimation: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
+  animationSpeed: PropTypes.number,
 }